#141 Changes after code review.
